backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyotp==2.8.0
qrcode[pil]==7.4.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML Libraries - FIXED VERSIONS for Python 3.12
numpy==1.26.0
pandas==2.1.4
scikit-learn==1.3.2
prophet==1.1.5
joblib==1.3.2

# PDF & Report Generation
reportlab==4.0.9
openpyxl==3.1.2
xlsxwriter==3.1.9

# Caching & Queues
redis==5.0.1
celery==5.3.4
apscheduler==3.10.4

# WebSocket & Real-time
python-socketio[asgi]==5.10.0
websockets==12.0

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2
pillow==10.1.0
aiofiles==23.2.1
setuptools  # Added for build compatibility
wheel      # Added for build compatibility